requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.6
Django==5.2.5
django-allauth==65.11.0
django-environ==0.12.0
django-filter==25.1
django-jazzmin==3.0.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.1
idna==3.10
jwt==1.4.0
kombu==5.5.4
Markdown==3.8.2
packaging==25.0
phonenumbers==9.0.12
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==6.4.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==12.5.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
